third commit

+++ src/controllers/company.controller.js
@@ -3,6 +3,8 @@ import { ApiResponse } from "../utils/ApiResponse.js";
 import { Company } from "../models/company.model.js";
 import { User } from "../models/user.model.js";
 import jwt from 'jsonwebtoken';
+import { Project } from "../models/project.model.js";
+import { Task } from "../models/task.model.js";
 
 const createCompany=async(req,res)=>{
     try {
@@ -74,8 +76,16 @@ const editCompany=async(req,res)=>{
 const deleteCompany=async(req,res)=>{
    try {
         const {companyId} = req.body ;
+
+        const comp = await Company.findById({companyId});
+        if(!comp){
+          return (res.json("Company not found"))
+        }
+        
+
         const deletingCompany = await Company.deleteOne({_id:companyId});
         const deleteProject = await Project.deleteMany({company:companyId});
+
         if (deleteProject.deletedCount === 0) {
           return res.status(404).json({ message: "No projects found in given Company" });
       }
